+++ test/badConfig.js
@@ -1,0 +1,4 @@
+module.exports = {
+  one: 1
+  two: 2
+};